capstone-project/requirements.txt

psycopg2-binary== 2.8.6
pandas==1.2.3
numpy==1.20.2
jupyter==1.0.0
ipython-sql==0.4.0
keyring== 23.0.1
boto3<1.16.0,>=1.15.0
botocore<1.19.0,>=1.18.18
colorama<0.4.4
docutils<0.16,>=0.10
PyYAML<5.4,>=3.10
flask-sqlalchemy==2.5.1
matplotlib==3.4.0
seaborn==0.11.1
celery==3.1.7
#pip install cloudpickle==1.4.1 distributed==2.17.0
## Data
requests==2.25.1
pathlib==1.0.1
zipfile36==0.1.3

## Sklearn & ML
scikit-learn==0.24.1
streamlit==0.79.0

## Airflow
apache-airflow==2.0.1
apache-airflow-providers-cncf-kubernetes==1.0.2
apache-airflow-providers-ftp==1.0.1
apache-airflow-providers-http==1.1.0
apache-airflow-providers-imap==1.0.1
apache-airflow-providers-postgres==1.0.1
apache-airflow-providers-sqlite==1.0.1
apache-airflow-upgrade-check==1.3.0
apache-airflow-providers-amazon==1.2.0

## Spark
pyspark
stop-words
clean-text